crewai/requirements.txt
crewai>=1.0.0
crewai-tools
langchain
langchain-openai
python-dotenv
